+++ src/main/java/umc/TripPiece/apiPayload/code/BaseCode.java
@@ -1,4 +1,4 @@
-package umc.TripPiece.payload.code;
+package umc.TripPiece.apiPayload.code;
 
 public interface BaseCode {
