+++ src/main/java/umc/TripPiece/apiPayload/exception/handler/EmojiHandler.java
@@ -1,0 +1,8 @@
+package umc.TripPiece.apiPayload.exception.handler;
+
+import umc.TripPiece.apiPayload.code.BaseErrorCode;
+import umc.TripPiece.apiPayload.exception.GeneralException;
+
+public class EmojiHandler extends GeneralException {
+    public EmojiHandler(BaseErrorCode errorCode) {super(errorCode);}
+}
