--- src/main/java/umc/TripPiece/payload/exception/handler/PictureHandler.java
@@ -1,8 +1,0 @@
-package umc.TripPiece.payload.exception.handler;
-
-import umc.TripPiece.payload.code.BaseErrorCode;
-import umc.TripPiece.payload.exception.GeneralException;
-
-public class PictureHandler extends GeneralException {
-    public PictureHandler(BaseErrorCode errorCode) {super(errorCode);}
-}
