--- src/main/java/umc/TripPiece/payload/exception/handler/VideoHandler.java
@@ -1,8 +1,0 @@
-package umc.TripPiece.payload.exception.handler;
-
-import umc.TripPiece.payload.code.BaseErrorCode;
-import umc.TripPiece.payload.exception.GeneralException;
-
-public class VideoHandler extends GeneralException {
-    public VideoHandler(BaseErrorCode errorCode) {super(errorCode);}
-}
